Upgrade to Websocket fix
